Add check in monitor to skip duplicate offline nodes with same IP
It is possible for the storage driver to return information for 2 nodes with the
same IP if a new node was re-using the IP from a node that was removed.
In that case the health monitor would incorrectly determine that the storage
driver was offline on the node.

This change removes offline nodes with duplicate IPs