PWX-22507 deploy apps on storageless nodes (#690)
* PWX-22507 run apps on storageless nodes

We want to be able to run sharedv4 (or any apps) on storageless nodes for automation. This PR adds a new job that cordon all the storage nodes and then schedule applications on top of the rest of (storage less) nodes.
When there's no storageless node, the test is skipped. 

Signed-off-by: dahuang <dahuang@purestorage.com>