PWX-22606: Give low scores to nodes which are in Degraded state.
- A node if in Degraded state will return half of the score they would
originally return based on the node/zone/rack/region priority.
- If a volume has a replica on a node but the node is in Degraded state
stork will give it a score of rackPriority / 2. This is done since
a pod running on that node is not really hyperconverged and should be
score similar to any another node in the same rack.
- Added UTs for scenarios where a node needs to be scored based on
zone/region/node/rack but is in Degraded state.