PWX-23093 Validates all app counter is active before going into failover (#735)
* Validates all app counter is active before going into failover

Reboot tests were sometimes failing because tests are not stable before going into the failover stage. This PR makes sure the number of active pods are the same as expected number of running pods before going into failover stage. When validating the terminated pods, we also validate the pods are shown as "not found".

Signed-off-by: dahuang <dahuang@purestorage.com>