PWX-24274: start nfs-server service during ReallocateSharedMount test (#806)
The test stops PX on a node and reboots the node. When the node comes
back up, NFS server is not running since we start NFS server only
after PX is ready. This can cause bonnie processes getting stuck
in D state on the client nodes and can lead to pods stuck in
terminating state.

This patch starts the NFS server before validating the context.