PWX-24637: mutate the virt-launcher container to intercept statfs()
Live migration of KubeVirt VM fails if the VM is using
a bind-mounted sharedv4 volume. The root cause is that libvirt
uses a statfs() call to check the file system type and
incorrectly concludes that the volume is not shared.

This patch addresses this problem as described below.

We use a shared library px_statfs.so that intercepts libvirt's statfs call.
If the input path is backed by a PX volume, we change the file system
type returned by the real statfs call. This shared library is bundled with
the stork container.

If a virt-launcher pod is being created and is using a PX volume, stork's
mutating webhook creates a ConfigMap in the pod's namespace.

This configMap has 2 keys that represent the 2 files that we want to inject
into the virt-launcher container's /etc directory:

  - ld.so.preload
  - px_statfs.so

The stork webhook then mutates the virt-launcher pod's spec to mount
the configMap as a volume and inject the 2 files above in /etc dir on the
container's file system. This makes linux load px_statfs.so in the libvirt
process that is running inside the virt-launcher container and intercept
libvirts' statfs() call.

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>