Add checks for readonly path (#815)
What this PR does / why we need it:
We were seeing cases where the readonly path was not set correctly, where both the device path and mount path are set to RO. This PR adds checks to make sure we detect it before proceeding further. However, this does not prevent the test from failing. We need a more robust way to mount the device for the expected permissions.

Which issue(s) this PR fixes (optional)
PWX-24555
Signed-off-by: dahuang <dahuang@purestorage.com>