PTX-10293: added tests for the webhook changes for virt-launcher pod
Tests to verify that the virt-launcher pod will return "nfs" as
the file system type for regular and encrypted PX volumes.

We use "kubevirt.io: virt-launcher" label to simulate the virt-launcher pod.

Also, verify that the pods without the virt-launcher label will return
the correct FS type depending on it is a bind-mount or an nfs-mount.

Signed-off-by: Neelesh Thakur <neelesh.thakur@purestorage.com>