PWX-36885: verify initial VMI state correctly
The tests were verifying the initial VMI only for the first VM in the list.
For the remaining VMs, the VMI has changed due to the testing performed on the first VM.
So, we end up verifying the second or third VMI. This caused a bug to escape.

Changed the tests to verify the initial VMI correctly and
verifed that they fail on OCP 4.14. Then, added an annotation to turn off
cdi population to make the tests pass.

Also, vendored in the latest sched-ops to handle empty Kind in GetPodsUsingPV.