PWX-36842: wait for kubevirt pvc to be bound
In OCP 4.14, stork receives the filter and prioritize requests while a temp PVC is being
cloned in another namespace. Since the real PVC is "Pending", stork is not able to assign
higher scores to the replica nodes for hyperconvergence.

Inside the filter request, we now detect that the PVC is not yet bound and return an error so that
k8s retries the filter request.